requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
faiss-cpu
sentence-transformers==2.2.2
google-generativeai==0.8.6
python-dotenv==1.0.0
langchain-text-splitters==0.0.1
pydantic==2.5.3

